requirements.txt
anyio==4.11.0
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==43.0.3
decorator==4.4.2
filelock==3.20.0
fsspec==2025.9.0
google-api-core==2.28.0
google-api-python-client==2.185.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.71.0
h11==0.14.0
httpcore==0.17.3
httplib2==0.31.0
httpx==0.24.1
idna==3.11
imageio==2.37.0
imageio-ffmpeg==0.6.0
Jinja2==3.1.6
llvmlite==0.45.1
MarkupSafe==3.0.3
more-itertools==10.8.0
moviepy==1.0.3
mpmath==1.3.0
networkx==3.5
numba==0.62.1
numpy==2.3.4
oauth2client==4.1.3
oauthlib==3.3.1
openai-whisper @ git+https://github.com/openai/whisper.git@c0d2f624c09dc18e709e37c2ad90c039a4eb72a2
packaging==25.0
pillow==11.3.0
proglog==0.1.12
proto-plus==1.26.1
protobuf==6.33.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
PyDrive2==1.21.3
pyOpenSSL==24.2.1
pyparsing==3.2.5
python-dotenv==1.2.1
python-telegram-bot==20.5
PyYAML==6.0.3
regex==2025.10.23
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
sympy==1.14.0
tiktoken==0.12.0
torch==2.9.0
tqdm==4.67.1
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
